Add Pipes Functionality and Add Pipes to the Game
Since we are going to handle a lot of pipes in the game, it's easier to
use a Phaser feature called 'group'. The group will contain 20 pipes
that we will be able to use as we want. So we add this in the create()
function.

we need a new function to add a pipe in the game. By default, all the
pipes contained in the group are dead and not displayed. So we pick a
dead pipe, display it, and make sure to automatically kill it when it's
no longer visible. This way we never run out of pipes.

The previous function displays one pipe, but we need to display 6 pipes
in a row with a hole somewhere in the middle. So let's create a new
function that does just that.

To actually add pipes in the game, we need to call the addRowOfPipes()
function every 1.5 seconds. We can do this by adding a timer in the
create() function.